+++ src/mps/mpo.jl
@@ -66,7 +66,7 @@ MPO(N::Int) = MPO(Vector{ITensor}(undef, N))
 Make an MPO with pairs of sites `s[i]` and `s[i]'`
 and operators `ops` on each site.
 """
-function MPO(::Type{ElT}, sites::Vector{<:Index}, ops::Vector{String}) where {ElT<:Number}
+function MPO(::Type{ElT}, sites::Vector{<:Index}, ops::Vector) where {ElT<:Number}
   N = length(sites)
   ampo = OpSum() + [ops[n] => n for n in 1:N]
   M = MPO(ampo, sites)
@@ -100,6 +100,15 @@ function MPO(::Type{ElT}, sites::Vector{<:Index}, op::String) where {ElT<:Number
 end
 
 MPO(sites::Vector{<:Index}, op::String) = MPO(Float64, sites, op)
+
+function MPO(::Type{ElT}, sites::Vector{<:Index}, op::Matrix{<:Number}) where {ElT<:Number}
+  # return MPO(ElT, sites, fill(op, length(sites)))
+  return error(
+    "Not defined on purpose because of potential ambiguity with `MPO(A::Array, sites::Vector)`. Pass the on-site matrices as functions like `MPO(sites, n -> [1 0; 0 1])` instead.",
+  )
+end
+
+MPO(sites::Vector{<:Index}, op::Matrix{ElT}) where {ElT<:Number} = MPO(ElT, sites, op)
 
 function randomMPO(sites::Vector{<:Index}, m::Int=1)
   M = MPO(sites, "Id")