+++ NEWS.md
@@ -6,8 +6,29 @@ Note that as of Julia v1.5, in order to see deprecation warnings you will need t
 
 After we release v1 of the package, we will start following [semantic versioning](https://semver.org).
 
+ITensors v0.3.12 Release Notes
+==============================
+
+Bugs:
+
+Enhancements:
+
+- Fix `svd` and `qr` for empty input left or right indices (#917)
+- Add support for defining MPOs from operators represented as matrices (#904)
+
+ITensors v0.3.11 Release Notes
+==============================
+
+Bugs:
+
+Enhancements:
+
+- Introduce `removeqn` function for removing a specified quantum number (#915)
+- Non-Hermitian `dmrg` (#913)
+- Clean up QN `svd` code in `ITensors` by handling QN blocks better in `NDTensors` (#906)
+
 ITensors v0.3.10 Release Notes
-=============================
+==============================
 
 Bugs: